Can now check for liars
entering ID or full name into terminal will now allow you to 100% know a customers real info. Later I will have a button to call them out if it is a payment discrepancy,  other than that it will be useful to know age for scamming old people ID for understanding they have a fake ticket and name for knowing they will be a higher risk to the park!